requirements.txt
python-dotenv
langchain
google-generativeai
streamlit